+++ subprojects/network-either/common/main/kotlin/com/javiersc/network/either/NetworkEither.kt
@@ -4,6 +4,7 @@ import arrow.core.Either
 import arrow.core.left
 import arrow.core.right
 import kotlin.contracts.contract
+import kotlin.jvm.JvmName
 
 public sealed class NetworkEither<out F, out S> {