+++ subprojects/network-either/common/main/kotlin/com/javiersc/network/either/internal/Constants.kt
@@ -1,7 +1,10 @@
 package com.javiersc.network.either.internal
 
+@Suppress("HttpUrlsUsage")
 internal object Constants {
-    const val DnsTimeout = 2000
-    const val DnsUrl = "https://8.8.8.8/"
-    const val DnsPort = 53
+
+    const val APPLE_URL = "http://captive.apple.com/hotspot-detect.htm"
+    const val GOOGLE_URL = "http://connectivitycheck.gstatic.com/generate_204"
+    const val PORT = 53
+    const val TIMEOUT = 2000
 }