+++ subprojects/network-either/jvm/test/kotlin/com/javiersc/network/either/_config/MockResponseUtil.kt
@@ -3,8 +3,8 @@ package com.javiersc.network.either._config
 import com.javiersc.network.either.Headers
 import org.mockserver.model.HttpResponse
 
-internal fun mockResponse(code: Int, file: String?, headers: Headers): HttpResponse =
+internal fun mockResponse(code: Int, content: String?, headers: Headers): HttpResponse =
     HttpResponse.response().withStatusCode(code).apply {
-        file?.let { withBody(readResource(it)) }
+        content?.let<String, HttpResponse?>(::withBody)
         headers.forEach { withHeader(it.key, it.value.first()) }
     }